refactor: make loan app a soft dependency
- add decorator to check if lending app is installed

- run salary slip loan functions only if lending app is installed
